+++ eslint-suppressions.json
@@ -9,7 +9,7 @@
   },
   "src/components/DeckRow.tsx": {
     "complexity": {
-      "count": 14
+      "count": 13
     },
     "max-lines": {
       "count": 1
@@ -115,7 +115,7 @@
   },
   "src/components/shelf/GameCard.tsx": {
     "complexity": {
-      "count": 7
+      "count": 6
     }
   },
   "src/components/shelf/HeroBackground.tsx": {
@@ -151,9 +151,6 @@
   "src/components/shelf/shelfStyles.ts": {
     "complexity": {
       "count": 9
-    },
-    "max-lines": {
-      "count": 1
     }
   },
   "src/components/styles/DeckModalStyles.tsx": {
@@ -311,7 +308,7 @@
   },
   "src/steam/index.ts": {
     "complexity": {
-      "count": 39
+      "count": 38
     },
     "max-lines": {
       "count": 1
