+++ src/runtime/deckyPlatform.ts
@@ -1,5 +1,5 @@
 import { Navigation } from "./host/decky";
-import { getAppMeta, getAppName, listCollections, listLibraryTabs, resolveShelfAppIds } from "../steam";
+import { getAppMeta, getAppMetaBatch, getAppName, listCollections, listLibraryTabs, resolveShelfAppIds } from "../steam";
 import type { ShelfSource } from "../types";
 import type { PlatformApi } from "./platform";
 import { logError } from "./logger";
@@ -81,6 +81,7 @@ export function createDeckyPlatform(): PlatformApi {
     resolveShelfAppIds,
     getAppName,
     getAppMeta,
+    getAppMetaBatch,
     navigateToApp: navigate,
     navigateToShelfSource,
   };
